refactor: restructure logging directories and server middleware
