+++ src/model/condition/ConditionFactory.ts
@@ -40,7 +40,7 @@ import {
     TimeConditionOptions,
     VariableConditionOptions
 } from '@p2olab/polaris-interface';
-import {catCondition} from '../../config/logging';
+import {catCondition} from '../../logging/logging';
 import {Module} from '../core/Module';
 import {AndCondition} from './AndCondition';
 import {Condition} from './Condition';
