+++ src/model/condition/NotCondition.ts
@@ -24,7 +24,7 @@
  */
 
 import {NotConditionOptions} from '@p2olab/polaris-interface';
-import {catCondition} from '../../config/logging';
+import {catCondition} from '../../logging/logging';
 import {Module} from '../core/Module';
 import {Condition} from './Condition';
 import {ConditionFactory} from './ConditionFactory';
