+++ src/model/condition/VariableCondition.ts
@@ -24,7 +24,7 @@
  */
 
 import {VariableConditionOptions} from '@p2olab/polaris-interface';
-import {catCondition} from '../../config/logging';
+import {catCondition} from '../../logging/logging';
 import {Module} from '../core/Module';
 import {Condition} from './Condition';
 import {ModuleCondition} from './ModuleCondition';
