+++ src/moduleTestServer/ModuleTestOpMode.ts
@@ -26,7 +26,7 @@
 import {Namespace, UAObject} from 'node-opcua-address-space';
 import {StatusCodes} from 'node-opcua-constants';
 import {DataType, Variant} from 'node-opcua-variant';
-import {catTestServer} from '../config/logging';
+import {catTestServer} from '../logging/logging';
 import {OpMode} from '../model/dataAssembly/mixins/OpMode';
 
 export class ModuleTestOpMode {
