+++ src/server/router/playerRouter.ts
@@ -25,7 +25,7 @@
 
 import {Request, Response, Router} from 'express';
 import * as asyncHandler from 'express-async-handler';
-import {catServer} from '../../config/logging';
+import {catServer} from '../../logging/logging';
 import {Manager} from '../../model/Manager';
 
 export const playerRouter: Router = Router();
